rag_system/flask_ui/frontend/requirements.txt
flask
db-dtypes
python-dotenv
